+++ src/components/tables/bookingsLists.tsx
@@ -22,7 +22,7 @@ export default function BookingsListTable({
     start_date: string;
     end_date: string;
   }>();
-  const [sort, setSort] = useState("asc");
+  const [sort, setSort] = useState("desc");
   const [currentPage, setCurrentPage] = useState(1);
 
   const { data, pagination, isLoading, isFailed, setIsFailed, retryFunction } =
@@ -30,7 +30,7 @@ export default function BookingsListTable({
       page: currentPage,
       start_date: filterDates?.start_date,
       end_date: filterDates?.end_date,
-      sort: sort,
+      sort,
       type,
     });