+++ src/pages/requests/requisition-requests/requisition-request-form.tsx
@@ -2,7 +2,6 @@ import { useCallback, useEffect, useState } from "react";
 import Switch from "../../../components/switch";
 import AddEditMaintenanceRequest from "../../../components/maintenance-requests/newRequest";
 import TextInput from "../../../components/inputs/textInput";
-import Select from "../../../components/inputs/select";
 import FileInputDesignTwo from "../../../components/inputs/fileInput/design-two/file-upload";
 import {
   requisitionRequestFormMain,
@@ -10,6 +9,9 @@ import {
 } from "../../../types/apiData/requisition-request";
 import purgeEmptyPayload from "../../../utils/remove-empty-payload";
 import useGetRequisitionRequest from "../../../services-hooks/userGetRequisitionRequest";
+import BankSearch from "../../../components/inputs/search/bank-search";
+import { Bank } from "../../../types/apiData/banks";
+import useVerifyBank from "../../../services-hooks/useVerifyBank";
 
 export default function RequisitionRequestForm({
   id, //maintenance_request_id
@@ -30,7 +32,7 @@ export default function RequisitionRequestForm({
 }) {
   const [markAsPaid, setMarkAsPaid] = useState(false);
   const [vendorName, setVendorName] = useState("");
-  const [vendorBank, setVendorBank] = useState("");
+  const [vendorBank, setVendorBank] = useState<Bank>({} as any);
   const [vendorAccountName, setVendorAccountName] = useState("");
   const [vendorAccountNumber, setVendorAccountNumber] = useState("");
   const [invoice, setInvoice] = useState<{
@@ -41,6 +43,10 @@ export default function RequisitionRequestForm({
     is_local: boolean;
   }>({} as any);
 
+  const { data: verfiedResult } = useVerifyBank({
+    bankCode: vendorBank?.code,
+    accountNumber: vendorAccountNumber,
+  });
   const { data } = useGetRequisitionRequest({ id: requisition_id });
   //   populate field provided id is available denoting update functionality
   useEffect(() => {
@@ -56,7 +62,7 @@ export default function RequisitionRequestForm({
       } = data;
       setMarkAsPaid(Boolean(is_paid));
       setVendorName(vendor_name || "");
-      setVendorBank(vendor_bank || "");
+      setVendorBank({ name: vendor_bank || "" } as any);
       setVendorAccountName(account_name || "");
       setVendorAccountNumber(account_number || "");
       setInvoice({
@@ -69,13 +75,20 @@ export default function RequisitionRequestForm({
     }
   }, [requisition_id, data]);
 
+  // POPULATE field on bank verification
+  useEffect(() => {
+    if (verfiedResult?.account_name) {
+      setVendorAccountName(verfiedResult?.account_name);
+    }
+  }, [verfiedResult]);
+
   const handleSubmit = useCallback(
     async (payload: requisitionRequestFormMain) => {
       try {
         const newPayload = {
           ...payload,
           vendor_name: vendorName,
-          vendor_bank: vendorBank,
+          vendor_bank: vendorBank?.name,
           account_name: vendorAccountName,
           account_number: vendorAccountNumber,
           invoice_file: invoice?.id ? "" : invoice,
@@ -127,7 +140,13 @@ export default function RequisitionRequestForm({
               inputType="text"
               label="Vendor's name"
             />
-            <Select
+            <BankSearch
+              placeholder="Search bank"
+              label="Vendor's Bank"
+              selected={vendorBank}
+              setSelected={setVendorBank}
+            />
+            {/* <Select
               isRequired={true}
               value={vendorBank}
               setValue={setVendorBank}
@@ -137,17 +156,12 @@ export default function RequisitionRequestForm({
               <option value="" disabled>
                 Select bank
               </option>
-              <option value="access">Access Bank</option>
-            </Select>
-            <TextInput
-              id="account-name"
-              placeholder="What's the account name?"
-              isRequired={true}
-              value={vendorAccountName}
-              setValue={setVendorAccountName}
-              inputType="text"
-              label="Account name"
-            />
+              {banks.map((item) => (
+                <option key={item?.id} value={item?.code}>
+                  {item?.name}
+                </option>
+              ))}
+            </Select> */}
             <TextInput
               id="account-number"
               placeholder="12345698700"
@@ -156,6 +170,16 @@ export default function RequisitionRequestForm({
               setValue={setVendorAccountNumber}
               inputType="number"
               label="Account number"
+            />
+            <TextInput
+              id="account-name"
+              placeholder="What's the account name?"
+              isRequired={true}
+              // readonly={true}
+              value={vendorAccountName}
+              setValue={setVendorAccountName}
+              inputType="text"
+              label="Account name"
             />
           </div>
           <div className="w-full flex flex-col gap-2">