+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ##### 22.05.2025
 
+> Upodate: Bookings overview reservations query orders changed to descending by default
 > Bug fix: Customers list pagination bug fixed
 > Updated: Dashboard analytics request breakdown url updated